refactor: Refactor services and implementations
- Moved interfaces to service package and implementations to a new
package
- Removed the old `impl` package
