+++ backend/src/main/java/com/chatop/rental_portal_backend/services/IAuthService.java
@@ -1,4 +1,4 @@
-package com.chatop.rental_portal_backend.services.impl;
+package com.chatop.rental_portal_backend.services;
 
 import com.chatop.rental_portal_backend.dto.AuthenticatedUserDTO;
 import com.chatop.rental_portal_backend.dto.LoginRequestDTO;
