+++ backend/src/main/java/com/chatop/rental_portal_backend/services/IMessageService.java
@@ -1,4 +1,4 @@
-package com.chatop.rental_portal_backend.services.impl;
+package com.chatop.rental_portal_backend.services;
 
 import com.chatop.rental_portal_backend.dto.MessageRequestDTO;
