+++ backend/src/main/java/com/chatop/rental_portal_backend/services/implementations/AuthService.java
@@ -1,4 +1,4 @@
-package com.chatop.rental_portal_backend.services;
+package com.chatop.rental_portal_backend.services.implementations;
 
 import java.util.Optional;
 
@@ -21,8 +21,8 @@ import com.chatop.rental_portal_backend.mappers.AuthenticatedUserMapper;
 import com.chatop.rental_portal_backend.mappers.RegisterUserMapper;
 import com.chatop.rental_portal_backend.models.User;
 import com.chatop.rental_portal_backend.repositories.UserRepository;
-import com.chatop.rental_portal_backend.services.impl.IAuthService;
-import com.chatop.rental_portal_backend.services.impl.IJwtService;
+import com.chatop.rental_portal_backend.services.IAuthService;
+import com.chatop.rental_portal_backend.services.IJwtService;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -53,6 +53,19 @@ public class AuthService implements IAuthService {
         this.passwordEncoder = passwordEncoder;
     }
 
+    /**
+     * Registers a new user in the system.
+     *
+     * This method first checks if a user with the provided email already exists.
+     * If the user exists, it throws a UserAlreadyExistsException. Otherwise, it
+     * maps the RegisterRequestDTO to a User entity, encodes the user's password,
+     * and saves the user to the repository. After successfully registering the user,
+     * it logs the registration event, authenticates the user, and generates a JWT token.
+     *
+     * @param registerRequestDTO the data transfer object containing the user's registration details
+     * @return a JWT token for the authenticated user
+     * @throws UserAlreadyExistsException if a user with the provided email already exists
+     */
     @Override
     public String registerUser(RegisterRequestDTO registerRequestDTO) {
         if (userRepository.findByEmail(registerRequestDTO.getEmail()).isPresent()) {
@@ -67,12 +80,31 @@ public class AuthService implements IAuthService {
         return generateToken(authentication);
     }
 
+    /**
+     * Logs in a user.
+     *
+     * This method authenticates the user with the provided email and password and
+     * generates a JWT token for the authenticated user.
+     *
+     * @param loginRequestDTO the data transfer object containing the user's login details
+     * @return a JWT token for the authenticated user
+     */
     @Override
     public String loginUser(LoginRequestDTO loginRequestDTO) {
         Authentication authentication = authenticateUser(loginRequestDTO.getEmail(), loginRequestDTO.getPassword());
         return generateToken(authentication);
     }
 
+    /**
+     * Retrieves the authenticated user.
+     *
+     * This method retrieves the authenticated user from the SecurityContext and maps
+     * the user entity to an AuthenticatedUserDTO. If no authenticated user is found,
+     * it throws a UserNotAuthenticatedException.
+     *
+     * @return an AuthenticatedUserDTO containing the authenticated user's details
+     * @throws UserNotAuthenticatedException if no authenticated user is found
+     */
     @Override
     public AuthenticatedUserDTO getAuthenticatedUser() {
         return getAuthenticatedUserFromContext()
@@ -83,6 +115,15 @@ public class AuthService implements IAuthService {
 
     }
 
+    /**
+     * Retrieves the authenticated user from the SecurityContext.
+     *
+     * This method retrieves the authenticated user from the SecurityContext and returns
+     * an Optional containing the user entity. If no authenticated user is found, it returns
+     * an empty Optional.
+     *
+     * @return an Optional containing the authenticated user
+     */
     private Optional<User> getAuthenticatedUserFromContext() {
         try {
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -99,6 +140,17 @@ public class AuthService implements IAuthService {
 
     }
 
+    /**
+     * Authenticates a user with the provided email and password.
+     *
+     * This method authenticates the user with the provided email and password using
+     * the AuthenticationManager interface.
+     *
+     * @param email the email of the user to authenticate
+     * @param password the password of the user to authenticate
+     * @return the authenticated user
+     * @throws InvalidLoginRequestException if the provided credentials are invalid
+     */
     private Authentication authenticateUser(String email, String password) {
         try {
             Authentication authentication = authenticationManager
@@ -110,6 +162,15 @@ public class AuthService implements IAuthService {
         }
     }
 
+    /**
+     * Generates a JWT token for the authenticated user.
+     *
+     * This method generates a JWT token for the authenticated user using the
+     * IJwtService interface.
+     *
+     * @param authentication the authenticated user
+     * @return a JWT token for the authenticated user
+     */
     private String generateToken(Authentication authentication) {
         return jwtService.generateToken(authentication);
     }
