+++ backend/src/main/java/com/chatop/rental_portal_backend/services/implementations/CloudinaryService.java
@@ -1,4 +1,4 @@
-package com.chatop.rental_portal_backend.services;
+package com.chatop.rental_portal_backend.services.implementations;
 
 import java.io.IOException;
 import java.util.Map;
@@ -7,7 +7,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.chatop.rental_portal_backend.services.impl.ICloudinaryService;
+import com.chatop.rental_portal_backend.services.ICloudinaryService;
 import com.cloudinary.Cloudinary;
 import com.cloudinary.utils.ObjectUtils;
 
@@ -20,6 +20,15 @@ public class CloudinaryService implements ICloudinaryService {
         this.cloudinary = cloudinary;
     }
 
+    /**
+     * Uploads an image to Cloudinary.
+     *
+     * This method uploads the provided image file to Cloudinary and returns the URL
+     * of the uploaded image.
+     *
+     * @param file the image file to upload.
+     * @return the URL of the uploaded image.
+     */
     @SuppressWarnings("unchecked")
     @Override
     public Optional<String> uploadImage(MultipartFile file) {
