+++ backend/src/main/java/com/chatop/rental_portal_backend/services/implementations/JwtService.java
@@ -1,4 +1,4 @@
-package com.chatop.rental_portal_backend.services;
+package com.chatop.rental_portal_backend.services.implementations;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
@@ -13,7 +13,7 @@ import org.springframework.security.oauth2.jwt.JwtEncoder;
 import org.springframework.security.oauth2.jwt.JwtEncoderParameters;
 import org.springframework.stereotype.Service;
 
-import com.chatop.rental_portal_backend.services.impl.IJwtService;
+import com.chatop.rental_portal_backend.services.IJwtService;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -29,6 +29,14 @@ public class JwtService implements IJwtService {
         this.jwtEncoder = jwtEncoder;
     }
 
+    /**
+     * Generates a JWT token.
+     * 
+     * This method generates a JWT token for the provided authentication object.
+     * 
+     * @param authentication the authentication object to generate the token for.
+     * @return the generated JWT token.
+     */
     @Override
     public String generateToken(Authentication authentication) {
         log.info("### Generating JWT token for user: {} ###", authentication.getName());
@@ -52,6 +60,14 @@ public class JwtService implements IJwtService {
         return jwtEncoder.encode(jwtEncoderParameters).getTokenValue();
     }
 
+    /**
+     * Decodes a JWT token.
+     * 
+     * This method decodes the provided JWT token and returns the decoded JWT object.
+     * 
+     * @param token the JWT token to decode.
+     * @return the decoded JWT object.
+     */
     @Override
     public Jwt decodeToken(String token) {
         return jwtDecoder.decode(token);
