+++ backend/src/main/java/com/chatop/rental_portal_backend/services/implementations/MessageService.java
@@ -1,4 +1,4 @@
-package com.chatop.rental_portal_backend.services;
+package com.chatop.rental_portal_backend.services.implementations;
 
 import org.springframework.stereotype.Service;
 
@@ -6,7 +6,7 @@ import com.chatop.rental_portal_backend.dto.MessageRequestDTO;
 import com.chatop.rental_portal_backend.mappers.SendMessageMapper;
 import com.chatop.rental_portal_backend.models.Message;
 import com.chatop.rental_portal_backend.repositories.MessageRepository;
-import com.chatop.rental_portal_backend.services.impl.IMessageService;
+import com.chatop.rental_portal_backend.services.IMessageService;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -22,6 +22,13 @@ public class MessageService implements IMessageService {
         this.sendMessageMapper = sendMessageMapper;
     }
 
+    /**
+     * Sends a message.
+     *
+     * This method saves the message to the database.
+     *
+     * @param messageRequest the message to send.
+     */
     @Override
     public void sendMessage(MessageRequestDTO messageRequest) {
         log.info(">>> MESSAGE SENT");
